fixing up weight download with tests
